keyboard layout: Base hidpi processing on gtk's scale state instead of letting QT5 do it itself.
On single-monitor machines, the keyboard layout was fine, but it
seems that QT determines scale based on the largest monitor (or at
least that's the issue that's manifesting here).  So if you have
one normal monitor, one 4k monitor, even if you're not in hidpi mode,
the keyboard layout is rendered twice as large (forcing the window
to be twice as wide also).

Instead, don't enable qt5 hidpi unless gtk says so, and pass this
along to the script.