partitioning: Handle exceptions differently for getDevice() and Disk()
This commit fixes a really bad bug.

If parted.getDevice() throws an exception:

- There's no need to format the device, we're dealing with something
which can't be read at device level here, not at partition table or
partition level.
- The disk_device variable is either NULL or worse, it still has
the value of the previous device. If the user answers YES to format the
device, it formats the previous one.

This bug is introduced by ad29db2ebf2a8a9e12302e3fe190f881efec4ca4 and
present in the LMDE 6 BETA.